Fix bug in _attempt_split in ProxGradSplit, refactor Hessian linops to remove code duplication.
