+++ rlaopt/splitting/linops.py
@@ -1,5 +1,8 @@
 """Helper linear operators for splitting representations."""
 
+from functools import partial
+from typing import Callable
+
 import torch
 from linops import LinearOperator
 
@@ -85,12 +88,87 @@ class _HVPLinOp(LinearOperator):
 
     def _matmul_impl(self, v: torch.Tensor) -> torch.Tensor:
         """Compute Hessian @ v using forward-over-reverse autodiff."""
+        return _pearlmutter_hvp(
+            lambda x: self._smooth_expr.evaluate(x), self._variable_values, v
+        )
 
-        def grad_dot_v(var_vals: TensorDict) -> torch.Tensor:
-            # Compute gradient of smooth_expr
-            grad = torch.func.grad(lambda x: self._smooth_expr.evaluate(x))(var_vals)
-            return torch.dot(grad.to_flat_tensor(), v)
 
-        # Differentiate grad_dot_v to get Hessian @ v
-        hvp_td = torch.func.grad(grad_dot_v)(self._variable_values)
-        return hvp_td.to_flat_tensor()
+class _SubampHVPLinOp(LinearOperator):
+    """Subsampled Hessian linear operator class.
+
+    Implements a linear operator interface for computing Hessian-vector products
+    using a subsampled batch of data. Uses forward-over-reverse automatic
+    differentiation for efficient computation.
+
+    Args:
+        loss (Callable): Loss function from which subsampled Hessian is constructed.
+            Should accept a TensorDict and return a scalar tensor.
+        variable_values (TensorDict): Variables at which the subsampled Hessian
+            is evaluated.
+        X_batch (torch.Tensor): Subsampled data matrix at which the loss is evaluated.
+        y_batch (torch.Tensor): Labels tensor at which the loss is evaluated.
+        device (torch.device): Device the model parameters live on.
+
+    Attributes:
+        device (torch.device): Device on which computations are performed.
+    """
+
+    def __init__(
+        self,
+        loss: Callable[[TensorDict, tuple[torch.Tensor, torch.Tensor]], torch.Tensor],
+        variable_values: TensorDict,
+        X_batch: torch.Tensor,
+        y_batch: torch.Tensor,
+        device: torch.device,
+    ):
+        """Initialize Subsampled Hessian linear operator.
+
+        Args:
+            loss (Callable): Loss function from which subsampled Hessian is constructed.
+                Should have signature loss(variable_values, X, y) -> scalar tensor.
+            variable_values (TensorDict): Variables at which the subsampled Hessian
+                is evaluated.
+            X_batch (torch.Tensor): Subsampled data matrix at which the loss is evaluated.
+            y_batch (torch.Tensor): Labels tensor at which the loss is evaluated.
+            device (torch.device): Device the model parameters live on.
+        """
+        super().__init__()
+        self._loss = partial(loss, X=X_batch, y=y_batch)
+        self._variable_values = variable_values
+
+        n = variable_values.flat_dim()
+        self._shape = (n, n)
+        self.device = device
+
+    def _matmul_impl(self, v: torch.Tensor) -> torch.Tensor:
+        """Compute Hessian-vector product using forward-over-reverse autodiff.
+
+        Computes the product of the subsampled Hessian matrix with a vector v
+        without explicitly forming the Hessian matrix. Uses automatic differentiation
+        to efficiently compute the Hessian-vector product.
+
+        Args:
+            v (torch.Tensor): Vector to multiply with the Hessian. Should have
+                shape (n,) where n is the flattened dimension of variable_values.
+
+        Returns:
+            torch.Tensor: The result of Hessian @ v as a flattened tensor.
+        """
+        return _pearlmutter_hvp(lambda x: self._loss(x), self._variable_values, v)
+
+
+def _pearlmutter_hvp(
+    f: Callable[[TensorDict], torch.Tensor],
+    variable_values: TensorDict,
+    v: torch.Tensor,
+) -> torch.Tensor:
+    """Computes Hessian vector product via Pearmutter's trick."""
+
+    def grad_dot_v(var_vals: TensorDict) -> torch.Tensor:
+        # Compute gradient of smooth_expr
+        grad = torch.func.grad(f)(var_vals)
+        return torch.dot(grad.to_flat_tensor(), v)
+
+    # Differentiate grad_dot_v to get Hessian @ v
+    hvp_td = torch.func.grad(grad_dot_v)(variable_values)
+    return hvp_td.to_flat_tensor()
